cli: change structure of the source file
split gobgp/main.go to main.go,global.go,neighbor.go,policy.go and common.go.
